fix: data type

+++ src/model/mvm/mvm_worker.cc
@@ -43,10 +43,10 @@ void MVMWorker::calculate_pctr(int start, int end){
   auto v = std::vector<float>();
   kv_v->Wait(kv_v->Pull(unique_keys, &v));
 
-  auto v_sum = std::vector<std::vector<int>>();
-  auto v_multi = std::vector<int>(end - start, 1.0);
+  auto v_sum = std::vector<std::vector<float>>();
+  auto v_multi = std::vector<float>(end - start, 1.0);
   for (int i = 0; i < sample_fgid_num.size(); ++i) {
-    auto fg_num = std::vector<int>(sample_fgid_num[i]);
+    auto fg_num = std::vector<float>(sample_fgid_num[i]);
     v_sum.push_back(fg_num);
   }
   for (size_t k = 0; k < v_dim_; ++k) {
@@ -115,7 +115,7 @@ void MVMWorker::calculate_gradient(std::vector<Base::sample_key>& all_keys,
     std::vector<ps::Key>& unique_keys,
     size_t start, size_t end,
     std::vector<float>& v,
-    std::vector<std::vector<int>>& v_sum,
+    std::vector<std::vector<float>>& v_sum,
     std::vector<float>& v_multi,
     std::vector<float>& loss,
     std::vector<float>& push_v_gradient) {
@@ -144,7 +144,7 @@ void MVMWorker::calculate_loss(std::vector<float>& v,
     std::vector<Base::sample_key>& all_keys,
     std::vector<ps::Key>& unique_keys,
     size_t start, size_t end,
-    std::vector<std::vector<int>>& v_sum,
+    std::vector<std::vector<float>>& v_sum,
     std::vector<float>& v_multi,
     std::vector<float>& loss) {
   for (size_t k = 0; k < v_dim_; k++) {
@@ -208,10 +208,10 @@ void MVMWorker::update(int start, int end){
   auto push_v_gradient = std::vector<float>(keys_size * v_dim_);
 
   auto loss = std::vector<float>(end - start);
-  auto v_sum = std::vector<std::vector<int>> ();
+  auto v_sum = std::vector<std::vector<float>> ();
   auto v_multi = std::vector<float>(end - start, 1.0);
   for (int i = 0; i < sample_fgid_num.size(); ++i) {
-    auto fg_num = std::vector<int>(sample_fgid_num[i]);
+    auto fg_num = std::vector<float>(sample_fgid_num[i]);
     v_sum.push_back(fg_num);
   }
   calculate_loss(v, all_keys, unique_keys, start, end, v_sum, v_multi, loss);
